IRvAdapterMutable extends IRvAdapter
